requirements.txt
aiohttp==3.7.4
asgiref==3.6.0
async-timeout==3.0.1
attrs==23.1.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.2
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
yarl==1.9.1
